Django_starter/requirement.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgi-redis==1.4.3
asgiref==3.3.1
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bbcode==1.1.0
bcrypt==3.2.0
boto==2.49.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
codecov==2.1.11
colorama==0.4.4
confusable-homoglyphs==3.2.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
distlib==0.3.1
Django==3.1.7
django-admin-views==0.8.0
django-adminlte2==0.4.3
django-adminplus==0.5
django-annoying==0.10.6
django-appconf==1.0.4
django-bootstrap-form==3.4
django-braces==1.14.0
django-bulk-update==2.2.0
django-chartit==0.2.9
django-chat==0.1.0
django-contrib-comments==2.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-eventstream==4.0.0
django-filter==2.4.0
django-frontend-template==1.2.0
django-grip==3.0.0
django-guardian==2.3.0
django-jasmine==0.5.1
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-oauth-toolkit==1.5.0
django-pagedown==2.2.0
django-private-chat==0.3.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-registration==3.1.1
django-registration-redux==2.9
django-selectable==1.3.0
django-simple-captcha==0.5.14
django-tables2==2.3.4
django-test-plus==1.4.0
django-threadedcomments==1.2
django-user-accounts==3.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jsonapi==4.1.0
djangorestframework-jwt==1.11.0
djappsettings==0.4.0
drf-json-api==0.1.1
filelock==3.0.12
future==0.18.2
futures==3.1.1
geolocation-python==0.2.2
gripcontrol==4.0.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.19.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jwcrypto==0.8
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mock==4.0.3
msgpack-python==0.5.6
mysqlclient==2.0.3
nose==1.3.7
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.7.2
pbr==5.5.1
Pillow==8.1.2
pinax-theme-bootstrap==8.0.1
pisa==3.0.33
pluggy==0.13.1
postmarkup==1.2.2
psycopg2==2.8.6
pubcontrol==3.0.0
py==1.10.0
pyasn1==0.4.8
pybbm==0.19.0
pycparser==2.20
PyJWT==2.0.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyPdf==1.13
PyPDF2==1.26.0
python-bidi==0.4.2
pytils==0.3
pytz==2021.1
RapidSMS==1.1.0
redis==3.5.3
reportlab==3.5.66
requests==2.25.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
storage==0.0.4.3
toml==0.10.2
tornado==6.1
tornado-psycopg2==0.3.1
tox==3.23.0
txaio==21.2.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
uuid==1.30
virtualenv==20.4.3
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
xhtml2pdf==0.2.5
xhtml2pdf-legacy==3.0.38
zope.interface==5.3.0
